+++ src/layouts/dpt10.js
@@ -3,9 +3,23 @@ module.exports = {
     base: {
         desc: "day of week + time of day",
         beforeDeserialize: function () {
-            return new Date(2000, 0, 1, 0, 0, 0);
+            return {};
         },
-        props: []
+        beforeSerialize: function (d) {
+            var m = d.match(/(\d+):(\d+):(\d+)/);
+            return {hours: m[1], minutes: m[2], seconds: m[3], day: 0};
+        },
+        props: [
+            {type: "uint", size: 6, index: "seconds"},
+            {type: "skip", size: 2},
+            {type: "uint", size: 6, index: "minutes"},
+            {type: "skip", size: 2},
+            {type: "uint", size: 5, index: "hours"},
+            {type: "uint", size: 3, index: "day"} // TODO: This value is currently ignored
+        ],
+        afterDeserialize: function (d) {
+            return `${d.hours}:${d.minutes}:${d.seconds}`;
+        }
     },
     subs: {
         // 10.001 time of day