+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.0.2 (18 nov 2016)
+
+- Now supports datapoints `DPT1` to `DPT19`.
+- Add overview of supported datapoints
+- Add method of loading custom property types
+- Start testing on individual datapoints
+
 ## 0.0.1 (16 nov 2016)
 
 First alpha. Supports datapoints `DPT1` to `DPT8` and `DPT12` to `DPT19`. Support for `DPT14` subtypes is limited.